+++ HISTORY.rst
@@ -13,6 +13,10 @@ Unreleased Changes
   indicating that the file passed validation. Previously, no output was emitted
   unless there were validation errors.
   https://github.com/natcap/geometamaker/issues/105
+* Added read support for a GDAL Raster Attribute Table. If it exists, the
+  table will be included as band metadata under the 'raster_attribute_table'
+  key. It can be retrieved by the ``get_rat`` method of a ``RasterResource``
+  instance. https://github.com/natcap/geometamaker/issues/25
 * The Natural Capital Project changed its name to the Natural Capital Alliance.
   References to the old name and website URL have been updated to reflect
   this change. https://github.com/natcap/geometamaker/issues/115